weather-dashboard/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
paho-mqtt>=1.6.1
python-dateutil>=2.9
